% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNZDM.R
\name{getNZDM}
\alias{getNZDM}
\title{Download Data From the NZDM Website.}
\usage{
getNZDM(
  x,
  destDir = NULL,
  quiet = TRUE,
  webAddress = "https://debtmanagement.treasury.govt.nz/investor-resources/data"
)
}
\arguments{
\item{x}{Name of dataset to retrieve. See details below.}

\item{destDir}{Directory into which the spreadsheet should be downloaded into.
Defaults to NULL in which case a temp directory is used.}

\item{quiet}{Logical passed to download.file.}

\item{webAddress}{URL where the spreadsheets are kept. Users don't need to change
this.}
}
\value{
A data frame (or a list of them if the spreadsheet contained multiple
        sheets) containing the requested data.
}
\description{
Retrieves data from
\url{https://debtmanagement.treasury.govt.nz/investor-resources/data}. This
involves downloading spreadsheets and reading them into R.
}
\details{
There are eleven datasets that can be retrieved whose names are listed
         below:

\itemize{
    \item Government bonds - tender issuance history
    \item Government Bonds - on issue (history at month end)
    \item Government Bonds - on issue (at month end)
    \item Government Bonds - Syndication
    \item Government bonds - tender history (repurchases)
    \item Treasury bills - tender history
    \item Treasury Bills - on issue (at month end)
    \item Treasury Bills - on issue (history at month end)
    \item European Commercial Paper -  on issue (at month end)
    \item Inflation-indexed bonds - factors spreadsheet
    \item Non market issuance - issuance history
}

There are a number of short-cuts used so that they are easier for the user to
specify. Check the code for exact details, but the gist of it is that spaces are
removed, letters are converted to lowercase, and certain words are truncated to
their initials. This means you can enter the required names in a variety of
formats (e.g. "Treasury Bill tender issuance history" is the same as
"tbtenderhistory" and hopefully they will be recognised. If in doubt you can
enter one of the names listed below.

\itemize{
    \item GB tender history
    \item GB onissue history
    \item GB onissue
    \item GB syndication
    \item GB repurchases
    \item TB tender history
    \item TB onissue
    \item TB onissue history
    \item ECP onissue
    \item IIB factors
    \item NMI history
}
}
\note{
The author has no affiliation with the NZDM office.
}
\examples{
\dontrun{
dat <- getNZDM('GB tender history')
plot(dat$Tender_Date, dat[, 'Wtd.Avg._successful_Yield_(\%)'], type = 'l')
}

}
\author{
Jasper Watson
}
